requirements/install.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/install.txt pyproject.toml
#
alembic==1.13.1
    # via bemserver-core (pyproject.toml)
amqp==5.2.0
    # via kombu
argon2-cffi==23.1.0
    # via bemserver-core (pyproject.toml)
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
async-timeout==4.0.3
    # via redis
billiard==4.2.0
    # via celery
celery==5.3.6
    # via bemserver-core (pyproject.toml)
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via
    #   argon2-cffi-bindings
    #   oso
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   bemserver-core (pyproject.toml)
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
greenlet==3.0.3
    # via sqlalchemy
idna==3.6
    # via requests
kombu==5.3.5
    # via celery
mako==1.3.2
    # via alembic
markupsafe==2.1.5
    # via mako
numpy==1.26.4
    # via pandas
oso==0.27.3
    # via bemserver-core (pyproject.toml)
pandas==2.2.0
    # via bemserver-core (pyproject.toml)
pint==0.23
    # via bemserver-core (pyproject.toml)
prompt-toolkit==3.0.43
    # via click-repl
psycopg==3.1.18
    # via bemserver-core (pyproject.toml)
pycparser==2.21
    # via cffi
python-dateutil==2.8.2
    # via
    #   celery
    #   pandas
pytz==2024.1
    # via pandas
redis==4.6.0
    # via bemserver-core (pyproject.toml)
requests==2.31.0
    # via bemserver-core (pyproject.toml)
six==1.16.0
    # via python-dateutil
sqlalchemy==2.0.26
    # via
    #   alembic
    #   bemserver-core (pyproject.toml)
typing-extensions==4.9.0
    # via
    #   alembic
    #   pint
    #   psycopg
    #   sqlalchemy
tzdata==2024.1
    # via
    #   celery
    #   pandas
urllib3==2.2.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
